+++ lib/src/ui/sections/interval_section.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:calendar_rrule_generator/src/ui/rrule_localizations.dart';
+import 'package:big_rrule_generator/src/ui/rrule_localizations.dart';
 
 /// Widget for selecting recurrence interval
 ///